use process::each_handle_exec_and_output
