+++ src/main.rs
@@ -1,6 +1,4 @@
-use std::process::{Command, Stdio};
 use std::sync::{Arc, Barrier, Mutex};
-use std::thread;
 
 mod log;
 mod output;
@@ -15,7 +13,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let scripts = procfile::scripts();
     let process_len = procfile::process_ln();
     let padding = procfile::padding();
-
     let barrier = Arc::new(Barrier::new(process_len + 1));
     let mut index = 0;
 
@@ -31,43 +28,19 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             let procs = procs.clone();
             let output = output.clone();
 
-            let handle_output = thread::Builder::new()
-                .name(String::from("handling output"))
-                .spawn(move || {
-                    let tmp_proc = process::Process {
-                        name: String::from(format!("{}.{}", key, n + 1)),
-                        child: Command::new(&script.cmd)
-                            .stdout(Stdio::piped())
-                            .stderr(Stdio::piped())
-                            .spawn()
-                            .unwrap(),
-                    };
-                    let proc = Arc::new(Mutex::new(tmp_proc));
-                    let proc2 = Arc::clone(&proc);
-
-                    let child_id = proc.lock().unwrap().child.id() as i32;
-                    output.log.output(
-                        "system",
-                        &format!("{0:1$} start at pid: {2}", &proc.lock().unwrap().name, padding, &child_id),
-                    );
-
-                    procs.lock().unwrap().push(proc);
-                    barrier.wait();
-
-                    output.handle_output(&proc2);
-                })?;
-
-            proc_handles.push(handle_output);
+            let each_fn = process::each_handle_exec_and_output(procs, padding, barrier, output);
+            let each_handle_exec_and_output = each_fn(String::from(key), n, String::from(&script.cmd));
+            proc_handles.push(each_handle_exec_and_output);
         }
     }
 
     barrier.wait();
 
     // use handle_signal
-    let procs_2 = Arc::clone(&procs);
+    let procs2 = Arc::clone(&procs);
     proc_handles.push(process::check_child_terminated(procs, padding));
 
-    let procs = Arc::clone(&procs_2);
+    let procs = Arc::clone(&procs2);
     proc_handles.push(signal::handle_signal(procs, padding));
 
     for handle in proc_handles {
